+++ wappstoiot/__init__.py
@@ -52,7 +52,7 @@ from .utils import name_check
 #                             __init__ Setup Stuff
 # #############################################################################
 
-__version__ = "v0.6.5"
+__version__ = "v0.6.6"
 __auther__ = "Seluxit A/S"
 
 __all__ = [
@@ -93,7 +93,6 @@ def onStatusChange(
     Configure an action when the Status have changed.
 
     def callback(StatusID: StatusID, newStatus: Any):
-
     """
     observer.subscribe(
         event_name=StatusID,
@@ -198,9 +197,7 @@ def _setup_IoTAPI(__config_folder, configs=None, fast_send=False):
 
 
 def _certificate_check(path) -> Dict[str, Path]:
-    """
-    Check if the right certificates are at the given path.
-    """
+    """Check if the right certificates are at the given path."""
     certi_path = {
         "ca": "ca.crt",
         "crt": "client.crt",
@@ -296,12 +293,20 @@ def createNetwork(
     name: str,
     description: str = "",
 ) -> Network:
+    """
+    Create a new Wappsto Network.
+
+    A Wappsto Network is references to the main grouping, of which multiple
+    device are connected.
+    """
     global __config_folder
     global __the_connection
 
-    if not name_check.legal_name(name):
+    illegal_chars: str = name_check.illegal_characters(name)
+
+    if illegal_chars:
         raise ValueError(
-            "Given name contain a ilegal character."
+            f"Given name contain a illegal character: {illegal_chars}"
             f"May only contain: {name_check.wappsto_letters}"
         )
 
@@ -329,15 +334,17 @@ def createNetwork(
 
 
 def connect():
+    """NOT Implemented yet."""
     pass
 
 
 def disconnect():
+    """NOT Implemented yet."""
     pass
 
 
 def close():
-    """."""
+    """Close down the connection to wappsto."""
     atexit.unregister(close)
     __ping_pong_thread_killed.set()
     __offline_storage_thread_killed.set()