+++ wappstoiot/modules/network.py
@@ -308,9 +308,11 @@ class Network(object):
         kwargs = locals()
         kwargs.pop('self')
 
-        if not name_check.legal_name(name):
+        illegal_chars: str = name_check.illegal_characters(name)
+
+        if illegal_chars:
             raise ValueError(
-                "Given name contain a ilegal character."
+                f"Given name contain a illegal character: {illegal_chars}"
                 f"May only contain: {name_check.wappsto_letters}"
             )